requirements.txt
python-dotenv
aiohttp
Jinja2
discord.py